requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
coverage==7.5.4
distro==1.9.0
Django==5.0.6
exceptiongroup==1.2.1
Faker==12.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
mixer==7.2.2
openai==1.35.10
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
